# docs/openapi/usuarios-service-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"tags":[{"name":"Perfiles internos","description":"Operaciones internas utilizadas por otros microservicios.\nEste controlador está destinado principalmente a la comunicación\nentre Auth Service y Usuarios Service.\n"}],"paths":{"/api/v1/tecnicos/{id}":{"get":{"tags":["tecnico-controller"],"operationId":"obtenerPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TecnicoResponse"}}}}}},"put":{"tags":["tecnico-controller"],"operationId":"actualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TecnicoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TecnicoResponse"}}}}}}},"/api/v1/perfiles/{id}":{"get":{"tags":["perfil-controller"],"operationId":"obtenerPorId_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilResponse"}}}}}},"put":{"tags":["perfil-controller"],"operationId":"actualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerfilUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilResponse"}}}}}},"delete":{"tags":["perfil-controller"],"operationId":"eliminar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/estudiantes/{id}":{"get":{"tags":["estudiante-controller"],"operationId":"obtenerPorId_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EstudianteResponse"}}}}}},"put":{"tags":["estudiante-controller"],"operationId":"actualizar_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstudianteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EstudianteResponse"}}}}}}},"/api/v1/docentes/{id}":{"get":{"tags":["docente-controller"],"operationId":"obtenerPorId_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocenteResponse"}}}}}},"put":{"tags":["docente-controller"],"operationId":"actualizar_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocenteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocenteResponse"}}}}}}},"/api/v1/administradores/{id}":{"get":{"tags":["administrador-controller"],"operationId":"obtenerPorId_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdministradorResponse"}}}}}},"put":{"tags":["administrador-controller"],"operationId":"actualizar_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdministradorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdministradorResponse"}}}}}}},"/api/v1/tecnicos":{"get":{"tags":["tecnico-controller"],"operationId":"listar","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTecnicoResponse"}}}}}},"post":{"tags":["tecnico-controller"],"operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TecnicoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TecnicoResponse"}}}}}}},"/api/v1/perfiles":{"get":{"tags":["perfil-controller"],"operationId":"listar_1","parameters":[{"name":"identificacion","in":"query","required":false,"schema":{"type":"string"}},{"name":"nombre","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"tipoPerfil","in":"query","required":false,"schema":{"type":"string","enum":["DOCENTE","ESTUDIANTE","TECNICO","ADMINISTRADOR"]}},{"name":"activo","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagePerfilResponse"}}}}}},"post":{"tags":["perfil-controller"],"operationId":"crear_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerfilCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilResponse"}}}}}}},"/api/v1/estudiantes":{"get":{"tags":["estudiante-controller"],"operationId":"listar_2","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageEstudianteResponse"}}}}}},"post":{"tags":["estudiante-controller"],"operationId":"crear_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstudianteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EstudianteResponse"}}}}}}},"/api/v1/docentes":{"get":{"tags":["docente-controller"],"operationId":"listar_3","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDocenteResponse"}}}}}},"post":{"tags":["docente-controller"],"operationId":"crear_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocenteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocenteResponse"}}}}}}},"/api/v1/administradores":{"get":{"tags":["administrador-controller"],"operationId":"listar_4","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAdministradorResponse"}}}}}},"post":{"tags":["administrador-controller"],"operationId":"crear_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdministradorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdministradorResponse"}}}}}}},"/api/v1/perfiles/{id}/estado":{"patch":{"tags":["perfil-controller"],"operationId":"cambiarEstado","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerfilEstadoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilResponse"}}}}}}},"/api/v1/internal/perfiles/{perfilId}/exists":{"get":{"tags":["Perfiles internos"],"summary":"Verificar la existencia de un perfil","description":"Permite que Auth Service compruebe si un perfil existe,\nsi está activo y qué tipos institucionales tiene asociados.\n\nLa solicitud debe incluir la cabecera X-Internal-Api-Key.\nEste endpoint no debe exponerse públicamente sin control\ndel Gateway o de la red interna.\n","operationId":"verificarExistenciaPerfil","parameters":[{"name":"perfilId","in":"path","description":"UUID del perfil que se desea validar","required":true,"schema":{"type":"string","format":"uuid"},"example":"6755fce4-9a44-48c5-9594-228e4667c036"},{"name":"X-Internal-Api-Key","in":"header","description":"Clave privada utilizada para autorizar la comunicación\nentre microservicios.\n","required":true,"schema":{"type":"string"},"example":"clave-interna-desarrollo"}],"responses":{"200":{"description":"La verificación fue realizada correctamente","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerfilExistsResponse"}}}},"400":{"description":"El perfilId no tiene un formato UUID válido"},"401":{"description":"La clave interna no fue enviada o es incorrecta","headers":{"WWW-Authenticate":{"description":"Indica que la solicitud interna no fue autorizada","style":"simple"}}},"500":{"description":"Ocurrió un error interno en el servidor"}}}},"/api/v1/docentes/perfil/{perfilId}":{"get":{"tags":["docente-controller"],"operationId":"obtenerPorPerfilId","parameters":[{"name":"perfilId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocenteResponse"}}}}}}}},"components":{"schemas":{"TecnicoRequest":{"type":"object","properties":{"perfilId":{"type":"string","format":"uuid"},"codigoTecnico":{"type":"string","maxLength":30,"minLength":0},"especialidad":{"type":"string","maxLength":100,"minLength":0},"activo":{"type":"boolean"}},"required":["codigoTecnico","perfilId"]},"TecnicoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"perfilId":{"type":"string","format":"uuid"},"codigoTecnico":{"type":"string"},"especialidad":{"type":"string"},"activo":{"type":"boolean"},"creadoEn":{"type":"string","format":"date-time"},"actualizadoEn":{"type":"string","format":"date-time"}}},"PerfilUpdateRequest":{"type":"object","properties":{"identificacion":{"type":"string","maxLength":20,"minLength":0},"nombres":{"type":"string","maxLength":100,"minLength":0},"apellidos":{"type":"string","maxLength":100,"minLength":0},"emailInstitucional":{"type":"string","maxLength":150,"minLength":0},"emailPersonal":{"type":"string","maxLength":150,"minLength":0},"telefono":{"type":"string","maxLength":20,"minLength":0},"direccion":{"type":"string","maxLength":255,"minLength":0},"fechaNacimiento":{"type":"string","format":"date"},"fotoUrl":{"type":"string"}},"required":["apellidos","emailInstitucional","nombres"]},"PerfilResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"identificacion":{"type":"string"},"nombres":{"type":"string"},"apellidos":{"type":"string"},"emailInstitucional":{"type":"string"},"emailPersonal":{"type":"string"},"telefono":{"type":"string"},"direccion":{"type":"string"},"fechaNacimiento":{"type":"string","format":"date"},"fotoUrl":{"type":"string"},"activo":{"type":"boolean"},"creadoEn":{"type":"string","format":"date-time"},"actualizadoEn":{"type":"string","format":"date-time"}}},"EstudianteRequest":{"type":"object","properties":{"perfilId":{"type":"string","format":"uuid"},"matricula":{"type":"string","maxLength":30,"minLength":0},"carreraId":{"type":"string","format":"uuid"},"semestre":{"type":"integer","format":"int32","maximum":20,"minimum":1},"activo":{"type":"boolean"}},"required":["matricula","perfilId"]},"EstudianteResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"perfilId":{"type":"string","format":"uuid"},"matricula":{"type":"string"},"carreraId":{"type":"string","format":"uuid"},"semestre":{"type":"integer","format":"int32"},"activo":{"type":"boolean"},"creadoEn":{"type":"string","format":"date-time"},"actualizadoEn":{"type":"string","format":"date-time"}}},"DocenteRequest":{"type":"object","properties":{"perfilId":{"type":"string","format":"uuid"},"codigoDocente":{"type":"string","maxLength":30,"minLength":0},"tituloAcademico":{"type":"string","maxLength":100,"minLength":0},"departamento":{"type":"string","maxLength":100,"minLength":0},"tipoContrato":{"type":"string","maxLength":30,"minLength":0},"dedicacion":{"type":"string","maxLength":30,"minLength":0},"activo":{"type":"boolean"}},"required":["perfilId"]},"DocenteResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"perfilId":{"type":"string","format":"uuid"},"codigoDocente":{"type":"string"},"tituloAcademico":{"type":"string"},"departamento":{"type":"string"},"tipoContrato":{"type":"string"},"dedicacion":{"type":"string"},"activo":{"type":"boolean"},"creadoEn":{"type":"string","format":"date-time"},"actualizadoEn":{"type":"string","format":"date-time"}}},"AdministradorRequest":{"type":"object","properties":{"perfilId":{"type":"string","format":"uuid"},"codigoAdministrador":{"type":"string","maxLength":30,"minLength":0},"cargo":{"type":"string","maxLength":100,"minLength":0},"pisoId":{"type":"string","format":"uuid"},"activo":{"type":"boolean"}},"required":["codigoAdministrador","perfilId"]},"AdministradorResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"perfilId":{"type":"string","format":"uuid"},"codigoAdministrador":{"type":"string"},"cargo":{"type":"string"},"pisoId":{"type":"string","format":"uuid"},"activo":{"type":"boolean"},"creadoEn":{"type":"string","format":"date-time"},"actualizadoEn":{"type":"string","format":"date-time"}}},"PerfilCreateRequest":{"type":"object","properties":{"identificacion":{"type":"string","maxLength":20,"minLength":0},"nombres":{"type":"string","maxLength":100,"minLength":0},"apellidos":{"type":"string","maxLength":100,"minLength":0},"emailInstitucional":{"type":"string","maxLength":150,"minLength":0},"emailPersonal":{"type":"string","maxLength":150,"minLength":0},"telefono":{"type":"string","maxLength":20,"minLength":0},"direccion":{"type":"string","maxLength":255,"minLength":0},"fechaNacimiento":{"type":"string","format":"date"},"fotoUrl":{"type":"string"}},"required":["apellidos","emailInstitucional","nombres"]},"PerfilEstadoRequest":{"type":"object","properties":{"activo":{"type":"boolean"}},"required":["activo"]},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageTecnicoResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TecnicoResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"sorted":{"type":"boolean"},"empty":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"PagePerfilResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/PerfilResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PerfilExistsResponse":{"type":"object","properties":{"perfilId":{"type":"string","format":"uuid"},"existe":{"type":"boolean"},"activo":{"type":"boolean"},"tiposPerfil":{"type":"array","items":{"type":"string"}}}},"PageEstudianteResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/EstudianteResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageDocenteResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DocenteResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageAdministradorResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AdministradorResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}